+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `onUpdate` didn't cancel `applyAction` and `invalidateAll`.
 - Hopefully fixed excessively deep infinity type instantiation on `message` helper.
+- Removed `errors.clear` optional parameter `undefinePath`, which was left there by mistake.
 
 ### Added